Bridge: Rename connection.rs -> session.rs.
