+++ src/bridge/session.rs
